core: Add a new XML-based serialisation format which can represent links
The GTK+ default serialisation format can't represent hyperlinks using
a GtkTextTag subclass as we were using. It's a better long-term move to have
our own serialisation format that we control anyway. This defines a new
data format version, and uses it as the default, converting entries as they
are edited/written.