core: Port to GMenu
This patch remove the main menubar from the main window replaced by a menu button in the 
toolbar following the new GNOME style. Rearranged the menu items, removing some of them, 
e.g. the print items (just "Print diary" in the GMenu).

Moved the code about the GMenu actions  from the main window to the application object.

Added a new .ui file for the GMenu because the Bug 667970 
https://bugzilla.gnome.org/show_bug.cgi?id=667970.

You can read about the menu reorganization discussion in Bug 676600 
https://bugzilla.gnome.org/show_bug.cgi?id=676600.