Bug 684412 - Add support for tagging diary entries
The main window now have an entry header, with the tag widget container
(AlmanahEntryTagsArea, base on a GtkGrid). This container is supposed to distribute the
tags (AlmanahTag) and the tag entry (AlmanahTagEntry) using the horizontal and vertical
space in a efficent way. It's responsible to load the tags for the selected entry.

The AlmanahTagEntry it's a GtkEntry with autocompletion capabilities with the tags available
in the diary storage.

AlmanahTag is a from scratch widget to draw the tag with some style using cairo.

The CSS has been updated to unify the visualization in the entry area.