App: Migrated GtkUIManager/GtkAction to GAction
All the stuff related to GtkUIManager and GActions has been removed to
support the modern GAction framework.

The actions in data/almana.ui has been added as main window actions,
with the "win." prefix, and keyboard accelerators has been moved to the
application.

The "jump-to-today" action has been removed, because the CalendarButton
can change directly the date in the Calendar.

https://bugzilla.gnome.org/show_bug.cgi?id=743096