VFS: Included a "finish" function
Because the Sqlite VFS requires a GSettings pointer
to read the encryption key ID, the sqlite3_vfs struct
is not a static struct but a kind of singleton.

So every time that you need to change the settings provider,
it's required to a call the almanah_vfs_finish function in
order to clear the last sqlite3_vfs struct (freeing memory)
and (the most important) unregister the Almanah VFS in the
SQLite3 list of VFS providers.

See https://bugzilla.gnome.org/show_bug.cgi?id=659500