vfs: Do not require secure memory for encryption/decryption buffers
Almanah is subject to RLIMIT_MEMLOCK, which limits the amount of memory
an unprivileged process can lock into physical memory. It’s typically
64KiB, which is a number we can easily hit if keeping the entire
decrypted database in memory.

Since keeping the database in memory (even if not locked into physical
memory) is better than storing it as a file on disk, fall back to
non-locked memory if the secure memory allocations fail. This means that
there is a chance the unencrypted database will hit the disk if swapped
out, but that’s something we’ll have to live with.

This additionally depends on the fix for gnome-keyring’s secure memory
allocators in https://bugzilla.gnome.org/show_bug.cgi?id=775685.