Add Nix based CI
I am using Nix for CI rather than the usual Fedora or Ubuntu,
since that is what I am most familiar with, and it also allows
to reproduce the same environment locally as on CI.

This includes the new CI-based publishing of release tarballs.
It will happen automatically in the `deploy` stage for signed tags
when the `build-and-test` job succeeds.