Replace cryptui with pure gpgme
libcryptui is unmaintained and archived. We already use gpgme for the cryptography so we might as well have it find the keys.

https://www.gnupg.org/documentation/manuals/gpgme/Listing-Keys.html
https://www.gnupg.org/documentation/manuals/gpgme/Key-objects.html

In preparation for GTK 4, let’s create a model object and use that as data source for a `GtkComboBox`. Once we get to the migration, we can just switch the model from extending `GtkListStore` to implementing `GListModel`.

This is slightly more verbose than libcryptui, as we have to replicate e.g. the id and label generation.

libcryptui also contains a (seahorse) daemon, which monitored the gnupg directory for updates and notified the apps when a new key appeared.
We now have to do the monitoring ourselves. But it should be fine since we can do it only when preferences are open.

It also required raising minimum gpgme version to 1.5.0 to be able to get the home directory easily.

https://www.gnupg.org/documentation/manuals/gpgme/Engine-Version-Check.html

And `g_string_free_and_steal` requires glib 2.76.

Fixes: https://gitlab.gnome.org/GNOME/almanah/-/issues/24