calendar-button: Use widget template
Slightly simplifies the code and required for GTK 4:
https://developer.gnome.org/documentation/tutorials/widget-templates.html

Had remove the `day-selected` signal in the .ui file as it started to fire on startup but since it comes from the template, it calls the signal handler `mw_calendar_day_selected_cb` with the wrong arguments (AlmanahCalendarButton*, AlmanahCalendar*).

Also had to rename some `id`s in the .ui file so that they can be used directly in `gtk_widget_class_bind_template_child`.

And removed stray `g_gobject_ref`. The objects are referenced by the template.